+++ packages/eyes-sdk-core/lib/EyesUtils.js
@@ -7,7 +7,7 @@ const {
   Region,
   MutableImage,
   EyesError,
-} = require('@applitools/eyes-common')
+} = require('..')
 const {EyesDriverOperationError} = require('./errors/EyesDriverOperationError')
 const EyesJsSnippets = require('./EyesJsSnippets')