+++ packages/eyes-sdk-core/lib/renderer/RenderStatusResults.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {GeneralUtils, RectangleSize, Region} = require('@applitools/eyes-common')
+const {GeneralUtils, RectangleSize, Region} = require('../..')
 
 /**
  * Encapsulates data for the render currently running in the client.