+++ packages/eyes-sdk-core/lib/wrappers/EyesWrappedElement.js
@@ -2,10 +2,10 @@
 const EyesUtils = require('../EyesUtils')
 
 /**
- * @typedef {import('@applitools/eyes-common').Logger} Logger
- * @typedef {import('@applitools/eyes-common').Region} Region
- * @typedef {import('@applitools/eyes-common').Location} Location
- * @typedef {import('@applitools/eyes-common').RectangleSize} RectangleSize
+ * @typedef {import('../logging/Logger').Logger} Logger
+ * @typedef {import('../geometry/Region').Region} Region
+ * @typedef {import('../geometry/Location').Location} Location
+ * @typedef {import('../geometry/RectangleSize').RectangleSize} RectangleSize
  * @typedef {import('./EyesWrappedDriver')} EyesWrappedDriver
  */