+++ packages/eyes-selenium/test/coverage/custom/TestEyesDifferentRunners.spec.js
@@ -177,13 +177,7 @@ async function validateVG(eyes) {
   let container = await eyes.getRunner().getAllTestResults(false)
   let results = container.getAllResults()
   for (let result of results) {
-    let data = await getApiData(
-      result
-        .getTestResults()
-        .getApiUrls()
-        .getSession(),
-      result.getTestResults().getSecretToken(),
-    )
+    let data = await getApiData(result.getTestResults())
     assert.deepStrictEqual(
       data.actualAppOutput.length,
       2,